`/ping` コマンドもcog_chat内に統合
`/ping` の機能がほぼ `/chat` 関連の機能に依存しているため
